Update bevy to 0.15

+++ src/common.rs
@@ -4,9 +4,9 @@ pub const CAMERA_SCALE: f32 = 0.25;
 pub const TILE_SIZE: f32 = 8.0;
 pub const PLAYER_GRAVITY_SCALE: f32 = 1.0;
 pub const PLAYER_DASHING_COLOR: Color = Color::srgb(
-    41 as f32 / u8::MAX as f32,
-    173 as f32 / u8::MAX as f32,
-    255 as f32 / u8::MAX as f32,
+    41f32 / u8::MAX as f32,
+    173f32 / u8::MAX as f32,
+    255f32 / u8::MAX as f32,
 );
 // 冲刺速度
 pub const PLAYER_DASH_SPEED: f32 = 200.0;
